Stop autoplaying if paused while navigating track list
Some of the default sounds at regular volumes are a little sharp on the
ear after some quiet ASMR so default to not playing a track if audio is
paused.

Since it appears the AppIndicator spec is lacking an action to prompt
metadata change without a play call —or sound_menu is lacking—, we
play & pause to give the UI a bump. This is awful. I know.

Fixes: https://bugs.launchpad.net/anoise/+bug/1784364