Correctly meet MPRIS2 spec for Metadata request
Noticed that according to
https://specifications.freedesktop.org/mpris-spec/latest/Track_List_Interface.html#Mapping:Metadata_Map
trackid was missing from the returned Metadata and it is required.

While doing that, noticed that VLC and other media players like to
include the URI for the track in the metadata, so added that too.

Picked /org/anoise/playlist/# for the dbus object path, but I know that
anoise.org is not for this app and … I strongly dislike the requirement
of reverse domain names for these things as it requires a domain or the
developer to just make something up — it’s why soo many Android apps use
com.gmail.username as their rdns