Retire ~/ANoise & ~/.ANoise data directories
Have to embrace the XDG current (not future).

Each time the track is changed, these locations are checked again, had
thought about building a dir list on launch and then never looking in
those locations again, ’cept with the Data Dir button addition to
preferences now making the location if it does not exist, we still want
to test the existence of that in case the user has added files at
runtime.

Until this application gains inotify support, cut out the redundancy.