Support watching relative sounds/ dir as well as DATA_DIR
This means that, when installed as a system tool,
/usr/share/anoise/sounds/ is watched for updates — when apt installs new
sound packs, they appear in the application straight away.

Further to this patch, we test directories exist before we watch them,
as watchdog doesn’t want to observe paths that do not exist.
It does force the user to need to restart the application if they add
DATA_DIR during first use, until we patch support to … runtime add newly
created dirs. Sigh.