Update the current noise index when watched dirs are changed
This means that the player now knows the new index in the tracklist for
the currently playing noise. It means that users will be less surprised
when they hit next and it could travel to the same current track.

There is one (pre)existing bug with this, with a user deleting the
active track, current index wont be updated, so when they navigate it
could jump by a few tracks. But it is acceptable. There’s a new bug
related to noises having been deleted during player restart, but since
config doesn’t track filename, and the later test will just set current
to 0, it doesn’t matter.

Now that the code wants to know the current track during __init__, the
max check is less important so remove it from _get_cfg_last()

Think this has been well tested by hand, but you never know.

Resolves #6