Add play toggle function
sound_menu has something similar, called PlayPause, which should
probably be updated to called _sound_menu_play_toggle() in case the
implementor wanted to do something specific to override this?

And while composing this message, I notice that the PlayPause method is
actually causing two signals to be announced for the same action.