Add toggling play function to sound_menu override options
As noticed in f6c14537b0b3a43571569a125a0ba00e2b7f6f87 we have a new
toggl‐ing play state function, and sound_menu already had a PlayPause
function that should be hooked up into it, so this does that.

It means that anoise could in the future do custom work in this toggling
state function. But eh, we wanted the feature for media keys anyhow.