More correctly meet media key behaviour
Keyboards such as the one in front of me have: ⏹ ⏮ ⏯ ⏭ as the only keys.
(Stop, Prev, Play/Pause, Next)

The Play/Pause key produces a singular Play key push, so assumption is
this key is a toggle and as such added support in
f6c14537b0b3a43571569a125a0ba00e2b7f6f87.
Since we didn’t yet have stop functionality, added that in
3f034efbca01ae68869966b14464b3e1c14d7b93 to reset track position.

This should be the end of working on support for this, but I expect
compat. issues to require quite a few updates.