Set friendly localised application name for Sound Preferences
Unlike our fixes with the icon and the general application genre for
pulseaudio, setting PULSE_PROP_application.name does not appear to
change at least the default behaviour in this version of Ubuntu. Our
application still shows as `anoise.py` to the client.

Documentation says this requires 2.2+ but since we are on a much later
version, and I did not see an obvious way to say ‘or later’ for the
requires, not included it in this patch.

However, other platforms may support that, so this is how it would be,
if it were enough:

```
os.environ[ 'PULSE_PROP_application.application_name' ] = "Ambient Noise"
os.environ[ 'PULSE_PROP_application.icon_name' ] = "anoise"
os.environ[ 'PULSE_PROP_media.role' ] = "music"
```

While spending time to discover how to do this, found that the majority
of these env vars are ‘loan words’ from libcanberra, so make a note in
the source.